% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{addListVectorEl}
\alias{addListVectorEl}
\title{Add element from list element's vector}
\usage{
addListVectorEl(stateList, transitionHead, transitionTail)
}
\arguments{
\item{stateList}{List of all the states storing the possible destination states}

\item{transitionHead}{State label which is the state where the transition is coming from}

\item{transitionTail}{State label which is the state where the transition is going to}
}
\description{
Add element from list element's vector
}
\keyword{internal}
