% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{cleanCohortData}
\alias{cleanCohortData}
\title{This function eliminates patients which do not fulfill the inclusion criteria}
\usage{
cleanCohortData(cohortData, mandatoryStates, outOfCohortAllowed = FALSE)
}
\arguments{
\item{cohortData}{Imported cohort data}

\item{mandatoryStates}{States which have to be present in the trajectory, otherwise dropped#'}
}
\value{
A dataframe with selected patients. Columns: cohort_definition_id, subject_id, cohort_start_date, cohort_end_date
}
\description{
This function eliminates patients which do not fulfill the inclusion criteria
}
\keyword{internal}
