\name{Cohort2Trajectory-package}
\alias{Cohort2Trajectory-package}
\alias{Cohort2Trajectory}
\docType{package}
\title{
  Package for creating patient orientated treatment trajectories from cohorts defined in OHDSI ATLAS
}
\description{
Package Cohort2Trajectory creates patient orientated treatment trajectories from cohorts defined in OHDSI ATLAS.
The package can be used with access to OMOP CDM database.
The package creates discrete and continuous time trajectories (and outputs them as .csv file) which describe patients' treatment through time.
Package can be run with GUI or CLI.
}
\details{
  This package can be used in two ways
  1) Running with GUI
  2) Running with CLI
}
\author{
Markus Haug, haugmarkus@protonmail.com.

Maintainer: Markus Haug haugmarkus@protonmail.com
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  HealthInformaticsUT/TrajectoryMarkovAnalysis,
  HealthInformaticsUT/TrajectoryViz
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
