% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{cohort2TrajectoryConfiguration}
\alias{cohort2TrajectoryConfiguration}
\title{C2T configuration}
\usage{
cohort2TrajectoryConfiguration(
  cdm = NULL,
  studyName = "Cohort2Trajectory",
  baseUrl = "http://localhost:8080/WebAPI",
  pathToStudy = getwd(),
  atlasTargetCohort = NULL,
  atlasStateCohorts = NULL,
  stateCohortLabels = NULL,
  stateCohortPriorityOrder = NULL,
  stateCohortMandatory = NULL,
  stateCohortAbsorbing = NULL,
  stateSelectionType = NULL,
  oocFix = "None",
  trajectoryType = "Discrete",
  lengthOfStay = NULL,
  outOfCohortAllowed = FALSE,
  runSavedStudy = FALSE,
  useCDM = TRUE,
  trajectoryDataObject = NULL,
  pathToData = "./Cohort2Trajectory/Data/importedData.csv",
  allowedStatesList = createStateList(stateCohortLabels),
  mergeStates = FALSE,
  mergeThreshold = 0.5,
  batchSize = 1000
)
}
\arguments{
\item{cdm}{object created with CDMConnector}

\item{studyName}{name chosen for study, is used as a folder name}

\item{baseUrl}{the base URL for the WebApi instance}

\item{pathToStudy}{path to folder which contains folder named after studyName}

\item{atlasTargetCohort}{the id of the target cohort defined in OHDSI tool ATLAS}

\item{atlasStateCohorts}{the ids of the state cohorts defined in OHDSI tool ATLAS}

\item{stateCohortLabels}{vector of the customized labels of the state cohorts}

\item{stateCohortPriorityOrder}{vector of the customized labels of the state cohorts in priority order}

\item{stateCohortMandatory}{vector of the customized labels of the state cohorts which are mandatory in trajectory}

\item{stateCohortAbsorbing}{vector of the customized labels of the state cohorts which are absorbing}

\item{stateSelectionType}{the type of state selection ("First" - First occurring, "Overlap" - Max overlap, "Priority" - Priority)}

\item{oocFix}{the method to use for replacing "OUT OF COHORT" states with more relevant states ("None" -> "OUT OF COHORT"; "Last present state" -> repeat the last one; random str -> used as state)}

\item{trajectoryType}{The type of the trajectory ("Discrete" - Discrete time, "Continuous" - Continuous time)}

\item{lengthOfStay}{The length of stay (days) in one state (Effect only in discrete case)}

\item{outOfCohortAllowed}{boolean whether the patient trajectory can surpass the target cohort's observation-period}

\item{runSavedStudy}{running a predefined study from studyName/Settings/trajectorySettings.csv}

\item{useCDM}{The package can also be run without the OMOP CDM}

\item{trajectoryDataObject}{When using without OMOP CDM specify the data file  (if specified no need for pathToData)}

\item{pathToData}{When using without OMOP CDM specify the path to data file (if specified no need for trajectoryDataObject)}

\item{allowedStatesList}{A list object which indicates accessible states from said state}

\item{mergeStates}{Boolean, if you want to merge states when they overlap}

\item{mergeThreshold}{Value from 0 to 1. If mergeStates is TRUE the states will be label-merged given they overlap more than the specified threshold. Can be given as vector, then multiple iterations are runned,}

\item{batchSize}{customizable batch size for cohort generation}
}
\description{
C2T configuration
}
\examples{
\dontrun{cohort2TrajectoryConfiguration(cdm = cdm,
studyName = "TestCohort2Trajectory",
baseUrl = NULL,
pathToStudy = getwd(),
atlasTargetCohort = "0",
atlasStateCohorts = c("test_state1", "test_state2"),
stateCohortLabels = c("test_label1", "test_label2"),
)}
}
