% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createMandatorySubDirs}
\alias{createMandatorySubDirs}
\title{Function which creates mandatory subdirectories and files to the pathToStudy directory}
\usage{
createMandatorySubDirs(pathToStudy, studyName)
}
\arguments{
\item{pathToStudy}{Path to the package}
}
\description{
Function which creates mandatory subdirectories and files to the pathToStudy directory
}
\keyword{internal}
