% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createStateList}
\alias{createStateList}
\title{Create a list of all states (as elements) for each state (as index)}
\usage{
createStateList(stateVector)
}
\arguments{
\item{stateVector}{Vector with the defines state labels}
}
\description{
Create a list of all states (as elements) for each state (as index)
}
\keyword{internal}
