% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryGeneration.R
\name{createTrajectories}
\alias{createTrajectories}
\title{Generate trajectories using the logic specified in the configuration}
\usage{
createTrajectories(
  cdm = NULL,
  studyEnv = NULL,
  trajectoryType = studyEnv$trajectoryType,
  runSavedStudy = studyEnv$runSavedStudy,
  oocFix = studyEnv$oocFix,
  outOfCohortAllowed = studyEnv$outOfCohortAllowed,
  lengthOfStay = studyEnv$lengthOfStay,
  stateCohortLabels = studyEnv$stateCohortLabels,
  stateCohortPriorityOrder = studyEnv$stateCohortPriorityOrder,
  stateSelectionType = studyEnv$stateSelectionType,
  stateCohortAbsorbing = studyEnv$stateCohortAbsorbing,
  stateCohortMandatory = studyEnv$stateCohortMandatory,
  allowedStatesList = studyEnv$allowedStatesList,
  useCDM = studyEnv$useCDM,
  pathToStudy = studyEnv$pathToStudy,
  batchSize = studyEnv$batchSize,
  updateAutomaticallyForMergedStates = FALSE
)
}
\arguments{
\item{cdm}{object created with CDMConnector}

\item{studyEnv}{environment created with cohort2TrajectoryConfiguration}

\item{trajectoryType}{The type of the trajectory ("Discrete" - Discrete time, "Continuous" - Continuous time)}

\item{runSavedStudy}{running a predefined study from studyName/Settings/trajectorySettings.csv}

\item{oocFix}{the method to use for replacing "OUT OF COHORT" states with more relevant states ("None" -> "OUT OF COHORT"; "Last present state" -> repeat the last one; random str -> used as state)}

\item{outOfCohortAllowed}{boolean whether the patient trajectory can surpass the target cohort's observation-period}

\item{lengthOfStay}{The length of stay (days) in one state (Effect only in discrete case)}

\item{stateCohortLabels}{vector of the customized labels of the state cohorts}

\item{stateCohortPriorityOrder}{vector of the customized labels of the state cohorts in priority order}

\item{stateSelectionType}{the type of state selection ("First" - First occurring, "Overlap" - Max overlap, "Priority" - Priority)}

\item{stateCohortAbsorbing}{vector of the customized labels of the state cohorts which are absorbing}

\item{stateCohortMandatory}{vector of the customized labels of the state cohorts which are mandatory in trajectory}

\item{allowedStatesList}{A list object which indicates accessible states from said state}

\item{useCDM}{The package can also be run without the OMOP CDM}

\item{pathToStudy}{path to directory that contains study folder}

\item{batchSize}{customizable batch size for trajectory generation process}
}
\value{
dataframe with trajectories
}
\description{
Generate trajectories using the logic specified in the configuration
}
\examples{
\dontrun{createTrajectories(cdm = cdm, runSavedStudy = F,studyEnv = studyEnv)}
}
