% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{generateCohortSetCSV}
\alias{generateCohortSetCSV}
\title{Generate cohort set CSV}
\usage{
generateCohortSetCSV(selectedStates, stateCohortLabels, pathToStudy, studyName)
}
\arguments{
\item{selectedStates}{all states to be included in db query}

\item{stateCohortLabels}{user inserted labels for non target states}

\item{pathToStudy}{path to folder that contains study folder}

\item{studyName}{name of the study directory}
}
\value{
cohortSetCSV df
}
\description{
internal function for generating cohort set CSV for CDMConnector generateCohorts
}
\keyword{internal}
