% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{getCohortData}
\alias{getCohortData}
\title{Get selected generated cohorts patient data}
\usage{
getCohortData(
  cdm,
  studyEnv,
  selectedTarget,
  selectedStates,
  baseUrl = "http://localhost:8080/WebAPI",
  pathToStudy = studyEnv$pathToStudy,
  studyName = studyEnv$studyName,
  stateCohortLabels = NULL
)
}
\arguments{
\item{cdm}{common data model reference object}

\item{studyEnv}{environment created with cohort2TrajectoryConfiguration}

\item{selectedTarget}{The sample cohort ID}

\item{selectedStates}{Vector of included cohort ID's}

\item{baseUrl}{the base URL for the WebApi instance}

\item{pathToStudy}{path to folder which contains folder named after studyName}

\item{studyName}{name chosen for study, is used as a folder name}

\item{stateCohortLabels}{Vector of the customized labels of the state cohorts}
}
\value{
A dataframe with selected patients. Columns: cohort_definition_id, subject_id, cohort_start_date, cohort_end_date
}
\description{
This function queries all patient ID's in selected cohorts
}
\keyword{internal}
