% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{getCohorts}
\alias{getCohorts}
\title{Get all generated cohort IDs}
\usage{
getCohorts(cdm)
}
\arguments{
\item{cdm}{common data model reference object}
}
\value{
A sorted vector of generated cohorts
}
\description{
This function queries all generated cohorts in Atlas
}
\keyword{internal}
