% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDataForStudy.R
\name{getDataForStudy}
\alias{getDataForStudy}
\title{Import the data from CDM for the study}
\usage{
getDataForStudy(
  runSavedStudy = studyEnv$runSavedStudy,
  useCDM = studyEnv$useCDM,
  trajectoryDataObject = studyEnv$trajectoryDataObject,
  cdm = NULL,
  studyEnv = NULL,
  atlasStateCohorts = studyEnv$atlasStateCohorts,
  atlasTargetCohort = studyEnv$atlasTargetCohort,
  stateCohortLabels = studyEnv$stateCohortLabels,
  baseUrl = studyEnv$baseUrl,
  outOfCohortAllowed = studyEnv$outOfCohortAllowed,
  stateCohortMandatory = studyEnv$stateCohortMandatory,
  stateCohortPriorityOrder = studyEnv$stateCohortPriorityOrder,
  mergeStates = studyEnv$mergeStates,
  mergeThreshold = studyEnv$mergeThreshold,
  allowedStatesList = studyEnv$allowedStatesList,
  stateCohortAbsorbing = studyEnv$stateCohortAbsorbing,
  pathToData = NULL
)
}
\arguments{
\item{runSavedStudy}{running a predefined study from studyName/Settings/trajectorySettings.csv}

\item{useCDM}{The package can also be run without the OMOP CDM}

\item{trajectoryDataObject}{When using without OMOP CDM specify the data file  (if specified no need for pathToData)}

\item{cdm}{object created with CDMConnector}

\item{studyEnv}{environment created with cohort2TrajectoryConfiguration}

\item{atlasStateCohorts}{the ids of the state cohorts defined in OHDSI tool ATLAS}

\item{atlasTargetCohort}{the id of the target cohort defined in OHDSI tool ATLAS}

\item{stateCohortLabels}{vector of the customized labels of the state cohorts}

\item{baseUrl}{the base URL for the WebApi instance}

\item{outOfCohortAllowed}{boolean whether the patient trajectory can surpass the target cohort's observation-period}

\item{stateCohortMandatory}{vector of the customized labels of the state cohorts which are mandatory in trajectory}

\item{stateCohortPriorityOrder}{vector of the customized labels of the state cohorts in priority order}

\item{mergeStates}{Boolean, if you want to merge states when they overlap}

\item{mergeThreshold}{Value from 0 to 1. If mergeStates is TRUE the states will be label-merged given they overlap more than the specified threshold. Can be given as vector, then multiple iterations are runned,}

\item{allowedStatesList}{List of allowed transitions between states}

\item{stateCohortAbsorbing}{Vector of absorbing states}

\item{pathToData}{When using without OMOP CDM specify the path to data file (if specified no need for trajectoryDataObject)}
}
\value{
log info
}
\description{
Import the data from CDM for the study
}
\examples{
\dontrun{getDataForStudy(cdm = cdm,studyEnv = studyEnv)}
}
