% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{getProfileData}
\alias{getProfileData}
\title{Get selected profile ID patient data}
\usage{
getProfileData(connection, dbms, cdmSchema, personId)
}
\arguments{
\item{connection}{Connection to database}

\item{dbms}{Database management system (postgresql, mysql etc.)}

\item{cdmSchema}{Name of the OMOP cdm schema}

\item{personId}{Person id value in person table}
}
\value{
A dataframe with selected patient's row data from OMOP database
}
\description{
This function queries patient's data from ohdsi_cdm person table
}
\keyword{internal}
