% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{getTrajectoriesContinuous}
\alias{getTrajectoriesContinuous}
\title{Prepare patientData dataframe for msm package}
\usage{
getTrajectoriesContinuous(
  cdm,
  patientData,
  stateSelection,
  statePriorityVector,
  absorbingStates = NULL,
  studyName = "",
  addPersonalData = TRUE,
  allowedStatesList = list()
)
}
\arguments{
\item{cdm}{a cdm object}

\item{patientData}{Object of class data.frame with columns subject_id, state_label, STATE_START_DATE, STATE_END_DATE}

\item{stateSelection}{Selection of the type of ordering}

\item{statePriorityVector}{All states in prioritized order}

\item{absorbingStates}{Absorbing states in dataframe}

\item{studyName}{name chosen for study, is used as a folder name}

\item{addPersonalData}{Logical, indicating whether or not to add personal data for each patient}

\item{allowedStatesList}{A list object which indicates accessible states from said state}
}
\value{
A dataframe ready for using msm package
}
\description{
This function prepares patientData dataframe for msm package
}
\keyword{internal}
