% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{getTrajectoriesDiscrete}
\alias{getTrajectoriesDiscrete}
\title{Convert the imported cohort data to patient treatmnet trajectories}
\usage{
getTrajectoriesDiscrete(
  cdm,
  cohortData,
  stateDuration = 30,
  stateSelection = 1,
  statePriorityVector = NULL,
  absorbingStates = NULL,
  oocFix = "None",
  studyName = "",
  addPersonalData = TRUE,
  allowedStatesList = list()
)
}
\arguments{
\item{cdm}{a cdm object}

\item{cohortData}{A data.frame object which is queried with getCohortsPatients function}

\item{stateDuration}{The length of a discrete state}

\item{stateSelection}{The chosen state selection method}

\item{statePriorityVector}{The order of priority for states}

\item{absorbingStates}{The absorbing states: patient trajectory ends with it}

\item{oocFix}{The method to use for replacing "OUT OF COHORT" states with more relevant states}

\item{studyName}{name chosen for study, is used as a folder name}

\item{addPersonalData}{Logical, indicating whether or not to add personal data for each patient}

\item{allowedStatesList}{A list object which indicates accessible states from said state}
}
\description{
This function outputs a data.frame object which describes the movement of patients between the defined states
}
\keyword{internal}
