% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{loadSettings}
\alias{loadSettings}
\title{Load settings of the study from trajectorySettings.csv according to the customized paramater studyName}
\usage{
loadSettings(studyName)
}
\arguments{
\item{studyName}{Customized name for the study}
}
\description{
Load settings of the study from trajectorySettings.csv according to the customized paramater studyName
}
\keyword{internal}
