% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{merge_overlaps}
\alias{merge_overlaps}
\title{Custom function to merge overlapping date ranges within each group}
\usage{
merge_overlaps(data)
}
\arguments{
\item{data}{A dataframe with raw cohort data}
}
\description{
Custom function to merge overlapping date ranges within each group
}
\keyword{internal}
