% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ordered_combinations}
\alias{ordered_combinations}
\title{Create a vector with all possible combinations preserving the priority order}
\usage{
ordered_combinations(states, n)
}
\arguments{
\item{states}{Vector of states}

\item{n}{The largest number of combinations possible}
}
\description{
Create a vector with all possible combinations preserving the priority order
}
\keyword{internal}
