% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{removeListVectorEl}
\alias{removeListVectorEl}
\title{Remove element from list element's vector}
\usage{
removeListVectorEl(stateList, transitionHead, transitionTail)
}
\arguments{
\item{stateList}{List of all the states storing the possible destination states}

\item{transitionHead}{State label which is the state where the transition is coming from}

\item{transitionTail}{State label which is the state where the transition is going to}
}
\description{
Remove element from list element's vector
}
\keyword{internal}
