% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{sanitize_filenames}
\alias{sanitize_filenames}
\title{Sanitize filenames not in correct format}
\usage{
sanitize_filenames(input_strings)
}
\arguments{
\item{input_strings}{A vector of state label names}
}
\description{
Sanitize filenames not in correct format
}
\keyword{internal}
