% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryGeneration.R
\name{updateMergedStates}
\alias{updateMergedStates}
\title{Update Cohort State Information for Merged States}
\usage{
updateMergedStates(
  data,
  stateCohortLabels,
  stateCohortPriorityOrder,
  allowedStatesList,
  stateCohortAbsorbing
)
}
\arguments{
\item{data}{A dataframe containing cohort definitions.}

\item{stateCohortLabels}{A vector of existing state labels.}

\item{stateCohortPriorityOrder}{A vector defining the priority order of states.}

\item{allowedStatesList}{A named list of allowed state transitions.}

\item{stateCohortAbsorbing}{A vector of absorbing states that prevent further transitions.}

\item{updateAutomaticallyForMergedStates}{A boolean flag to enable automatic updates.}
}
\value{
A list containing updated stateCohortLabels, stateCohortPriorityOrder, allowedStatesList, and stateCohortAbsorbing.
}
\description{
Updates stateCohortLabels, stateCohortPriorityOrder, allowedStatesList, and stateCohortAbsorbing
when new merged states are introduced.
}
\keyword{internal}
