% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisationFunctions.R
\name{visualisePatient}
\alias{visualisePatient}
\title{Visualize patient's trajectory pathway}
\usage{
visualisePatient(patientData, patientId, trajectoryStopDays = 183, theme = 1)
}
\arguments{
\item{patientData}{Object of class data.frame with columns SUBJECT_ID, STATE, STATE_START_DATE, STATE_END_DATE}

\item{patientId}{Patient's id}

\item{trajectoryStopDays}{Number of days which will determine the possible split between trajectories}

\item{theme}{Plot's theme indicator}
}
\description{
This function outputs a patient oriented plot
}
\keyword{internal}
