% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisationFunctions.R
\name{visualiseStateOverlap}
\alias{visualiseStateOverlap}
\title{Visualise overlaping cohorts to detect possible state's which would be advisible to join}
\usage{
visualiseStateOverlap(patientData, stateLabels)
}
\arguments{
\item{patientData}{Object of class data.frame with columns SUBJECT_ID, STATE_LABEL, STATE_START_DATE, STATE_END_DATE}

\item{stateLabels}{Labels for state notation}
}
\description{
This function outputs a pheatmap
}
\keyword{internal}
