% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{getCohorts}
\alias{getCohorts}
\title{Get all generated cohort IDs}
\usage{
getCohorts(connection, dbms, resultsSchema)
}
\arguments{
\item{connection}{Connection to database}

\item{dbms}{Database management system: postgresql, mysql etc.}

\item{resultsSchema}{Name of the OMOP results schema}
}
\value{
A sorted vector of generated cohorts
}
\description{
This function queries all generated cohorts in Atlas
}
\keyword{internal}
