% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{getTrajectoriesDiscrete}
\alias{getTrajectoriesDiscrete}
\title{Convert the imported cohort data to patient treatmnet trajectories}
\usage{
getTrajectoriesDiscrete(
  connection,
  dbms,
  cohortData,
  stateDuration = 30,
  pathToResults = getwd(),
  stateSelection = 1,
  statePriorityVector = NULL,
  absorbingStates = NULL,
  oocFix = "None",
  studyName = "",
  addPersonalData = TRUE,
  allowedStatesList = list()
)
}
\arguments{
\item{cohortData}{A data.frame object which is queried with getCohortsPatients function}

\item{stateDuration}{The length of a discrete state}

\item{pathToResults}{Path to temp folder for saving objects}

\item{stateSelection}{The chosen state selection method}

\item{statePriorityVector}{The order of priority for states}

\item{absorbingStates}{The absorbing states: patient trajectory ends with it}

\item{addPersonalData}{Logical, indicating whether or not to add personal data for each patient}
}
\description{
This function outputs a data.frame object which describes the movement of patients between the defined states
}
\keyword{internal}
